cheri: Fix compile in drivers/iommu/
Standard uintptr_t vs. unsigned long issues.

+++ drivers/iommu/iommu-pages.h
@@ -20,7 +20,7 @@ struct ioptdesc {
 	unsigned long __page_flags;
 
 	struct list_head iopt_freelist_elm;
-	unsigned long __page_mapping;
+	uintptr_t __page_mapping;
 	union {
 		u8 incoherent;
 		pgoff_t __index;
@@ -30,7 +30,7 @@ struct ioptdesc {
 	unsigned int __page_type;
 	atomic_t __page_refcount;
 #ifdef CONFIG_MEMCG
-	unsigned long memcg_data;
+	uintptr_t memcg_data;
 #endif
 };
